*: refactor to have SecureJoinVFS
The purpose of this API (and the VFS interface) is two-fold. The first
reason is that it allows us to mock-test all of our code, and the second
is that it allows external users to apply SecureJoin in cases where os.*
is not valid to use (this may include VFS-like structures or cases as in
umoci where CAP_DAC_OVERRIDE is emulated).

Signed-off-by: Aleksa Sarai <asarai@suse.de>