add go.mod and vendor upstream pkg/errors v0.8.0
This patch replaces the vendor.conf with a go.mod, and re-vendors
the pkg/errors files from upstream.

This commit is mostly to preserve history of the local changes that
were made to the pkg/errors package (which appear to have been made
for debugging purposes).

Setting the minimum go version in go.mod to v1.7 for this commit,
which matches the lowest version that's testet against in CI.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>