join: don't allow .. and . in working path during resolution
The previous implementation would allow lexical components like ".." to
be present in the intermediate path and wait for them to be cleaned at
the very end.

While this is technically okay to do (since we expand symlinks and the
components are definitely purely lexical when we add them to the working
path), it makes more sense to update the working path to equal the
generated path (which is cleaned). This also reduces the possibility of
a subtle bug here causing issues.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>