initial safe MkdirAll and MkdirAllHandle implementations
A fairly common usage of SecureJoin is of the form:

  path, _ := securejoin.SecureJoin(root, nonExistentPath)
  _ = os.MkdirAll(path, 0o755)

Unfortunately, such usage is not safe if the root can be modified
between SecureJoin and MkdirAll. This is a known issue with SecureJoin
(using path strings for the API means we cannot be sure the path is
safe).

However, by carefully using file handles it is possible to adapt
SecureJoin to be safe to use in a race-free way with MkdirAll by merging
them into a single function. The implementation is designed so that we
can easily add openat2 support, which is done in a follow-up commit.

MkdirAllHandle is a variant of MkdirAll which takes the root as an
*os.File and returns an *os.File for the final part of the created
directory tree. This is a preferable interface to use if you need a
handle to the final directory.

We have to update the minimum Go version to 1.20 so that we can have
errors that contain more than one wrapped error.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>