procfs: use /proc/thread-self to check
Rather than saving a handle to /proc/self/fd, save a /proc handle and
use /proc/thread-self/fd. This is necessary to handle Go programs with
locked threads that have unshare(CLONE_FS), such as runc.

This is modelled on similar logic that exists in runc (ProcThreadSelf),
and is necessary if we want to use MkdirAll in runc.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>