mkdirall: add tests for MkdirAll and MkdirAllHandle
These are primarily correctness tests (82.4% code coverage), to make
sure we handle all of the ugly corner-cases with nested and dangling
symlinks correctly.

Almost all of the untested code is related to errors (most of which are
annoying to mock). The one set of error paths it might be worth adding
errors for is the isDeadInode() checks.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>