mkdirall: make partialOpen semantically identical to openat2 implementation
openat2(2) will fail when we try to follow a broken symlink, which means
that partialOpenat2() will treat the final symlink as a non-existent
path. In contrast, partialOpen() will happily return a handle and
remainingPath in the middle of the symlink resolution.

While either of the above semantics are acceptable, we need to unify the
behaviour so that users don't get different behaviour based on their
kernel version. We can't change the behaviour of openat2(2), and
emulating the current partialOpen() behaviour using openat2(2) is just
asking for trouble. So, make partialOpen() match the behaviour of
partialOpenat2().

Unfortunately, implementing this properly (including handling recursive
symlinks) requires emulating a symlink stack and making sure that we
return the top symlink we are in the mille of resolving if we hit a
non-existent or otherwise invalid path. If there is no last symlink, we
return the partial open as normal. Yeah, this is a little ugly, but I
don't see a nicer way of implementing these semantics...

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>