procfs: add tests for our safe /proc helpers
This ensures we test all of the fallbacks even if we don't use them in
practice when doing MkdirAll in our test suite. This raises the test
coverage to 82.0%.

We are still missing tests for some of the auto-detection logic, and the
pre-3.17 fallback for /proc/thread-self. We can probably mock these out
the same way we do with openat2, but it might be a little annoying to
keep the code clean.

In terms of correctness tests, there are a couple of extra tests it's
probably worth adding (see procfs_linux_test.go for a list).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>