procfs: add tests for our safe /proc helpers
This ensures we test all of the fallbacks even if we don't use them in
practice when doing MkdirAll in our test suite. In addition, add some
hooks to allow the tests to force the usage of fallbacks when using the
standard getProcRoot() function everything else uses. This raises the
test coverage to 87.1%.

In terms of correctness tests, there are a couple of extra tests it's
probably worth adding (see procfs_linux_test.go for a list).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>