gha: collate and calculate coverage for all test runs
By combining the coverage from all platforms as well as privileged and
non-privileged Linux runs, the coverage is now 87.7% for the entire
project.

It was necessary to split the Windows CI scripts because mktemp doesn't
work in Powershell and it's necessary to reimplement the whole thing to
get what you need (not to mention that -test.gocoverdir needs to be
quoted in a particular way to not make Powershell angry).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>