tests: add racing tests for partialLokupInRoot and MkdirAllHandle
In order to get some confidence in saying that our code safe against
these kinds of racing attacks, add some tests to cover this.

The full suite (with 50k runs of each new subtest) takes about 10
minutes, so we should use "go test -short" for the full matrix and add a
single separate test run which can run for longer (we don't need to
merge the coverage from the stress test because we hit the same
codepaths as in the short test).

While working on this, I found what seems to be a kernel bug with
readlink(/proc/self/fd/$n). There is a workaround in the tests for this,
but it is possible that the test will become flaky.

This brings the test coverage up to 91.1% on Linux (91.3% combined).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>