lookup: use readlinkat(fd, "") to get link components
readlinkat(2) implies AT_EMPTY_PATH, so we can just use it to look up an
opened symlink component in a race-free way. This removes one possible
race (which we handled already, but it's nice to completely avoid it).

This feature was added in 2011 in commit 65cfc6722361 ("readlinkat(),
fchownat() and fstatat() with empty relative pathnames"), which is more
than old enough for us to depend on. If someone uses filepath-securejoin
on a v2.6.38 kernel, that's their own problem.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>