tests: fix race when pausing rename swap
The pauseCh was only synchronised one way, which resulted in a race
window here the test thread (after requesting the pause) would get the
real path before the rename thread swapped the file back. This is easily
fixed by doing the swap twice each loop iteration so that we only
receive pause requests when we are in an okay-to-be-paused state.

Removing the retry logic lets us do far more test runs for the racing
tests, removing the need for the -short suite. The MkdirAll tests are
still a bit slow, but 2k runs should be fine.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>