lookup: handle // and trailing slash components correctly
When we hit empty components, we need to treat them as though they are a
"." component. We could skip them for non-trailing components but for
trailing components it's critical to actually try to do the open so that
we get openat2-like errors for non-directory states.

For the single trailing slash case, it's simpler to implement it as a
final "." open after we've done the other lookups (we could switch to
using an array of path component like libpathrs does, but this is a
simpler change).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>