mkdirall: explicitly return an error for suid/sgid bits
These bits are ignored by mkdirat(2) on every POSIX system (Linux does
honour the sticky bit, though it is an outlier). We could ignore them
but this seems less than ideal (a user could be under the impression
their code works when actually the bit is not getting set) so returning
an informative error is preferable.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>