tests: don't panic if the fd is closed
Despite 1446241e8a9f ("tests: make sure the directory handle is kept
alive in rename loop"), we are still getting the same errors. I suspect
the issue is that the directory is getting deleted while we're still in
the loop. To work around this, don't panic if the directory is closed
(which must happen before the Go test infrastructure deletes the
temporary directory).

Fixes:1446241e8a9f ("tests: make sure the directory handle is kept alive in rename loop")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>