Isolate the testing import in test code
testing_mocks_linux.go imports the testing package, and that import
leaks into any binary built with securejoin. This causes issues in
projects which check for that, such as Kubernetes (see
https://github.com/kubernetes/kubernetes/pull/127832 and its tests).

This adds a level of indirection for hooks used to force testing
behaviour in procfs_linux.go, so that the mocks can be defined only in
_test.go files. This ensures that the testing package is only pulled
in for tests.

Signed-off-by: Stephen Kitt <skitt@redhat.com>