Fix compilation on 32bit architectures such as i386, armhf
On those architectures, the tests fail to build with an overflow error:

src/github.com/cyphar/filepath-securejoin/mkdir_linux_test.go:254:8: constant 2147484159 overflows int
FAIL	github.com/cyphar/filepath-securejoin [build failed]

Turns out that FileMode is defined as an uint32 in the os.fs package.
Since `MkdirAll` is intended as a replacement for os.MkdirAll, it makes
sense to use the same types for all arguments, and that means os.FileMode
for the mode parameter.

Signed-off-by: Reinhard Tartler <siretart@tauware.de>