tests: don't call testing.Testing() in mocks
With commit 686491226941 ("Isolate the testing import in test code"),
the testing hooks were only ever called in binaries compiled in test
mode, meaning that checking if we are in a test doesn't really make
sense.

This is also necessary to remove because testing.Testing() was added in
Go 1.21, which is too new for some of our users.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>