gha: test older Go versions in CI
In particular, we want to test our minimum (Go 1.18), as well as the
versions where have explicit compatibility handling (Go 1.20 and 1.21)
in order to make sure that builds with any mix of compatibility shims
work correctly.

One slight wrinkle is that -test.gocoverdir was only added in Go 1.20,
so for older Go versions we need to just skip generating coverage (in
theory you can collate coverage runs from older Go versions but the
format is different to the post-1.20 format and so would be too
complicated to deal with).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>