deps: downgrade golang.org/x/sys requirement
The newest feature from golang.org/x/sys we aboslutely need is the
fsconfig(2) wrappers which were added in v0.18.0. We also would like the
definition of STATX_MNT_ID_UNIQUE but since it is a fixed constant, we
don't need to bump the dependency all the way up to v0.20.0.

Note that this is being done to avoid forcing downstreams to upgrade
their golang.org/x/sys dependency unnecessarily. This does not mean that
all users of filepath-securejoin will use an outdated version -- Go only
uses a single version of any module for compiling entire programs, and
its use of Minimal Version Selection[1] means that if a user of
filepath-securejoin requires a newer version they will get the newer
version (and filepath-securejoin would use the same version too).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>