join: loosen cleanliness requirements for SecureJoin root
It turns out that some users do provide unclean paths like "foo/bar/"
and as a result the new behaviour in commit bc750ad62e1e ("join: return
an error if root is unclean path") was far too aggressive and lead to
regressions.

The more gentle solution is to only error out if the path contains a
".." component (which is the only component type we are really worried
about here because it's the only one that can turn a safe
root-joined-path into an unsafe one due to how symlinks are resolved on
Linux).

Fixes: bc750ad62e1e ("join: return an error if root is unclean path")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>