*: unify nil dir handling for *at(2) wrappers
There is a fair bit of boilerplate when dealing with nil (dir *os.File)
arguments, which we can easily simplify.

This also lets us avoid nil dereferences if some code accidentally
passes nil, and instead we will get EBADF if the path was relative but
the passed dirfd is nil. Most programs would default to AT_FDCWD, but
-EBADF is more preferable for us (to make sure we never accidentally
reference something relative to cwd by mistake).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>