proc: switch to much safer resolver for non-openat2 systems
Previously we would just use openat(2) to open subpaths with the "safe"
procfs resolver. For some users (namely, those that couldn't get a
non-overmounted handle and didn't have openat2), this was actually just
as unsafe as operating on /proc directly.

This problem has been solved in libpathrs (by creating a fairly minimal
resolver just for procfs paths, that removes the need for procfs-based
verification and also adds some additional protections regular
filesystem resolvers don't need), but due to the needs of that project
the resolver is a bit more complicated. For filepath-securejoin (which
only supports O_PATH|O_NOFOLLOW for resolution, followed by reopening)
we can use an even simpler lookup implementation.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>