*: relicense libpathrs ports to MPLv2
For historical reasons, the SecureJoin code needs to be BSD licensed, as
it is based on work written by folks from Docker (which originally based
it on the Go stdlib).

However, for all of the new API code (which is based on libpathrs), it
makes little sense to use such a permissive license (given the license
of libpathrs is copyleft). MPLv2 lets us have file-based copyleft
without affecting the (logically separated) old API code.

I am the sole author of the vast majority of this code, but we have had
some contributors that have made some changes to this code (and we thus
need their approval to be on the safe side when it comes to
relicensing):

 * Stephen Kitt wrote 1 minor testing-related patch.
 * Kir Kolyshkin wrote 3 minor documentation patches, and 2 patches
   related to moving to sync.OnceValue.

The following contributors have made contributions to this repository,
but not to any of the files which are being relicensed, and thus we do
not need their approval:

 * Jakub Wilk wrote 1 minor non-code documentation patch.
 * Paulo Gomes wrote 1 patch fixing a security issue in SecureJoin.
 * Devendranath Thadi wrote 1 non-code patch related to CI.

Signed-off-by: Stephen Kitt <skitt@redhat.com>
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>