procfs: cache whether subset=pid is supported
It turns out that the error path from fsconfig(2) is actually noticeably
slow, purely because string errors need to be formatted and stored in
buffers. Caching whether subset=pid is supported will help with this a
lot, and may obviate the need for caching fsopen(2) -- it turns out that
fsopen(2) and fsmount(2) are surprisingly cheap. AlmaLinux/RHEL 8 has
backported fsopen(2) but not the subset=pid stuff -- it's not clear if
this is the cause of some performance issues, but it won't hurt.

Note that we don't try to fallback to hidepid=2 or similar -- before
subset=pids was merged, each pidns had one procfs superblock and so
modifying this in runc would be a bad idea as it would impact the host
procfs mount (in practice it would be ignored by the kernel's mount
option merging logic, but best not to tempt fate).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>