procfs: rework API to be more handle-based
It turns out that while most of the new mount API is fairly cheap to
use, FSCONFIG_CMD_CREATE is somewhat expensive if you plan to replace a
lot of common high-frequency operations with it.

Caching the handles (as removed in commit 79161464a1b4 ("procfs: remove
caching of procfs root") and will be added in a later patch in this
series) helps with this a fair bit, but for operations that cannot be
scoped with subset=pid handles (such as writing sysctls) we need to
provide a way for users to take FSCONFIG_CMD_CREATE out of the hot path
at their own discretion.

Also, this more closely matches the current Rust API for libpathrs (and
the future planned C and Go APIs for the libpathrs wrappers).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>