internal: add kernelversion package
This was originally intended to be a simple copy of runc's internal
kernelversion package (which was derived from containerd, in turn
derived from Docker). However, the runc one is quite over-complicated
and uses quite confusing terminology and so I went ahead and wrote a
more generic but less complicated version.

It is very loosely based on the Go stdlib's kernelversion but completely
rewritten and extended to support more than 2-digit kernel versions (for
2.6.x support) and with quite a few added tests.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>