*: use internal Fd-like interface instead of *os.File
This is necessary in order to be able to have a cached global *os.File
that we can be sure will not have Close() called on it accidentally.

At the moment, this is a very minimal interface (most of the current
*os.File users internally only really care about the underlying file
descriptor), but it is very trivial to add more io.Fooer entries to the
interface.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>