procfs: re-add caching of subset=pid ProcfsHandle
This is a partial revert of commit 79161464a1b4 ("procfs: remove caching
of procfs root"), but with the following key differences:

 * We only cache the handle if we are sure it has subset=pid active,
   which significantly reduces the risks of it being leaked or having a
   long lifetime.

 * Previously, we just cached a basic *os.File which meant that an
   errant Close could result in spurious errors. Now that we have
   fd.NopCloser, we can disarm Close for the global cached handle.

The main reason for this revert is that while it is undoubtedly more
secure to never cache the handle, it turns out that while most of the
new mount API is fairly cheap to use, it is still quite a bit more
expensive than whatever operation you are going to be replacing (which
is very noticeable in high-frequency operations).

Ultimately, subset=pid is intended to be safe for these kinds of
purposes so while I would prefer to not have these kinds of leaks it is
probably okay in the grand scheme of things.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>