COPYING: add new COPYING file for pkg.go.dev
pkg.go.dev's scanner is fairly limited when it comes to scanning for
licenses in slightly more complicated projects. I much prefer SPDX
license tags in each file (along with a human-readable license header)
but the Go tooling only accepts certain files with certain names.

Unfortunately, LICENSE.BSD and LICENSE.MPL-2.0 are not on that list.
This patch is an unfortunate workaround by adding a textual description
of the license in a "known good" filename (along with the full license
text in the hopes of getting the pkg.go.dev scanner to accept it).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>