testutils: move test helpers to internal subpackage
This is needed for the future pathrs-lite subpackage split.

The only really complicated bit was dealing with WithWithoutOpenat2 --
we want to call t.Run but because it involves a concrete *testing.T type
we can't just do it the straightforward way. I ended up with a bit of
closure soup (passing around a TestingT interface) but hopefully it's
still fairly easy to reason about.

Effectively, we provide WithWithoutOpenat2 with a closure that wraps
t.Run so it can use it, with some extra function passing added so that
WithWithoutOpenat2 can do setup and call the func(*testing.T)-style
closure.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>