pathrs-lite: move new API to subpackage
This code should probably have been in a subpackage from the start
(since it has very little to do with securejoin) but at the time it felt
like a very minimal workaround.

Now that it looks like this is going to be a more permanent solution for
part of the Go ecosystem (and the license for most of this code has been
changed) we should separate it. The name "pathrs-lite" indicates what
the purpose of this subpackage is (i.e., it is a less complete version
of libpathrs).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>