internal/linux: do not cache HasOpenat2 success
Programs like runc apply seccomp filters to themselves during execution,
and so it is possible that subsequent calls into pathrs-lite could
result in errors because pathrs-lite thinks that openat2(2) is permitted
but the new seccomp filter denies it. This would result in ENOSYS (or
other errors) being returned from the openat2(2) resolver rather than
using the fallback.

This does add an extra openat2(2) syscall to every lookup, which is less
than ideal, but the alternatives are not much better:

 * We could always do the fallback if the openat2(2) lookup fails.

   This seems like the obvious choice from the outset but I didn't want
   to provide a mechanism for attackers to downgrade us to non-openat2
   lookups (if an attacker finds a bug in the opath resolver they cannot
   currently exploit it on openat2(2)-enabled systems, but if it causes
   openat2(2) to error out then always doing the fallback will reduce
   the overall security of libpathrs / pathrs-lite).

   We could do this only for -ENOSYS (which shouldn't be possible to
   trigger by an attacker) but some people use -EPERM for seccomp
   profiles as the default errno (or even explicitly use it) so this
   wouldn't solve the issue completely.

 * We could have a set-once cached value whenver we see an error in
   HasOpenat2 but the fallback code does so many syscalls it's probably
   not worth it.

 * Since this issue can only really happen due to seccomp filters being
   installed, a somewhat wacky idea would be to look at
   /proc/thread-self/status to see if Seccomp_filters has changed. This
   is obviously more expensive than just trying openat2(2).

As far as I can see, this issue doesn't apply to HasNewMountAPI because
getProcRoot() will always fallback to unsafeHostProcRoot() in case of
*any* error. This isn't the case for openat2, and is what causes the
issue.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>
(cherry picked from commit 28c869d6afc73fe9f4521ea796901e394c0d3335)
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>