gopathrs: try openat2 first before checking HasOpenat2
The most common case is that openat2(2) is supported and just works, so
we should just try to use it first when doing lookups. If that fails, we
can then check if openat2(2) in general doesn't work (falling back to
the O_PATH resolver) or if this was an error from that specific lookup
(in which case we return the error to avoid downgrade attacks).

This has the benefit of making the common case as fast as possible.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>